refactor(frontend): extract comfyui connection hook
